remove reconnect nonce for now, move ws utils
