+++ exchange/src/adapter/hub/hyperliquid/stream.rs
@@ -1,7 +1,7 @@
 use crate::{
     Event, Kline, Price, PushFrequency, TickMultiplier, Ticker, TickerInfo, Timeframe, Trade,
     Volume,
-    adapter::connect::{WsTransport, channel, connect_ws},
+    adapter::connect::{WsAdapter, WsControlConfig, WsTransport, channel, connect_ws},
     adapter::{MarketKind, StreamKind, StreamTicksize, TRADE_BUCKET_INTERVAL, flush_trade_buffers},
     depth::{DeOrder, DepthPayload, DepthUpdate, LocalDepthCache},
     serde_util::de_string_to_number,
@@ -9,10 +9,7 @@ use crate::{
 };
 
 use super::{HyperliquidHandle, WS_DOMAIN, raw_qty_unit_from_market_type};
-use crate::adapter::hub::{
-    AdapterError,
-    ws_control::{WsAdapter, WsControlConfig},
-};
+use crate::adapter::hub::AdapterError;
 use fastwebsockets::Frame;
 use futures::{SinkExt, Stream, channel::mpsc};
 use rustc_hash::FxHashMap;
